Refactor code to create separate folders for worker store and utils
