+++ packages/extension/src/devtools/components/tabs/CookieTab/cookieList.tsx
@@ -13,12 +13,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+/**
+ * External dependencies.
+ */
 import React from 'react';
 
 /**
  * Internal dependencies.
  */
-import { type CookieData } from '../../../../cookieStore';
+import { type CookieData } from '../../../../store';
 import ListItem from './listItem';
 
 interface ICoookieList {
