+++ packages/extension/src/store/consts.ts
@@ -13,25 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-/**
- * External dependencies.
- */
-import React from 'react';
-import { createRoot } from 'react-dom/client';
-
 /**
  * Internal dependencies.
  */
-import App from './app';
-import { Provider as ExternalStoreProvider } from '../app/cookieStore';
+import { type StorageValue } from './types';
 
-const root = document.getElementById('root');
-
-if (root) {
-  createRoot(root).render(
-    <ExternalStoreProvider>
-      <App />
-    </ExternalStoreProvider>
-  );
-}
+export const emptyTabData: StorageValue = {
+  cookies: {},
+  url: undefined,
+  focusedAt: undefined,
+};
