+++ packages/extension/src/utils/getCurrentTabId.ts
@@ -13,25 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+export const getCurrentTabId = async () => {
+  const tabs = await chrome.tabs.query({
+    active: true,
+    currentWindow: true,
+  });
 
-/**
- * External dependencies.
- */
-import React from 'react';
-import { createRoot } from 'react-dom/client';
-
-/**
- * Internal dependencies.
- */
-import App from './app';
-import { Provider as ExternalStoreProvider } from '../app/cookieStore';
-
-const root = document.getElementById('root');
-
-if (root) {
-  createRoot(root).render(
-    <ExternalStoreProvider>
-      <App />
-    </ExternalStoreProvider>
-  );
-}
+  return tabs[0].id;
+};
